requirements.txt
connexion
atomicwrites==1.3.0
attrs==19.3.0
certifi==2019.9.11
colorama==0.4.1
coverage==4.5.4
importlib-metadata==0.23
more-itertools==7.2.0
packaging==19.2
pipenv==2018.11.26
pluggy==0.13.0
py==1.8.0
pyparsing==2.4.2
pytest==5.2.1
pytest-cov==2.8.1
PyYAML==5.1.2
six==1.12.0
virtualenv==16.7.5
virtualenv-clone==0.5.3
wcwidth==0.1.7
zipp==0.6.0
flask_script
flask_migrate
flask_sqlalchemy
flask_marshmallow
marshmallow-sqlalchemy